move search related classes to a dedicated package to prepare scope limitation of most classes
