+++ protocols/imap/src/main/scala/org/apache/james/imap/search/SearchOperation.scala
@@ -1,0 +1,5 @@
+package org.apache.james.imap.search
+
+import org.apache.james.imap.api.message.request.SearchResultOption
+
+case class SearchOperation(key: SearchKey, options: Seq[SearchResultOption])
