+++ protocols/imap/src/test/java/org/apache/james/imap/search/SearchProcessorTest.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.imap.processor;
+package org.apache.james.imap.search;
 
 import static org.apache.james.imap.ImapFixture.TAG;
 import static org.mockito.ArgumentMatchers.eq;
@@ -44,52 +44,13 @@ import org.apache.james.imap.api.ImapConstants;
 import org.apache.james.imap.api.display.HumanReadableText;
 import org.apache.james.imap.api.message.IdRange;
 import org.apache.james.imap.api.message.UidRange;
-import org.apache.james.imap.api.message.request.All;
-import org.apache.james.imap.api.message.request.And;
-import org.apache.james.imap.api.message.request.Answered;
-import org.apache.james.imap.api.message.request.Bcc;
-import org.apache.james.imap.api.message.request.Body;
-import org.apache.james.imap.api.message.request.Cc;
 import org.apache.james.imap.api.message.request.DayMonthYear;
-import org.apache.james.imap.api.message.request.Deleted;
-import org.apache.james.imap.api.message.request.Draft;
-import org.apache.james.imap.api.message.request.Flagged;
-import org.apache.james.imap.api.message.request.From;
-import org.apache.james.imap.api.message.request.Header;
-import org.apache.james.imap.api.message.request.Keyword;
-import org.apache.james.imap.api.message.request.Larger;
-import org.apache.james.imap.api.message.request.New;
-import org.apache.james.imap.api.message.request.Not;
-import org.apache.james.imap.api.message.request.Old;
-import org.apache.james.imap.api.message.request.On;
-import org.apache.james.imap.api.message.request.Or;
-import org.apache.james.imap.api.message.request.Recent;
-import org.apache.james.imap.api.message.request.SearchKey;
-import org.apache.james.imap.api.message.request.SearchOperation;
 import org.apache.james.imap.api.message.request.SearchResultOption;
-import org.apache.james.imap.api.message.request.Seen;
-import org.apache.james.imap.api.message.request.SentBefore;
-import org.apache.james.imap.api.message.request.SentOn;
-import org.apache.james.imap.api.message.request.SentSince;
-import org.apache.james.imap.api.message.request.SequenceNumbers;
-import org.apache.james.imap.api.message.request.Since;
-import org.apache.james.imap.api.message.request.Smaller;
-import org.apache.james.imap.api.message.request.Subject;
-import org.apache.james.imap.api.message.request.Text;
-import org.apache.james.imap.api.message.request.To;
-import org.apache.james.imap.api.message.request.Uid;
-import org.apache.james.imap.api.message.request.UnAnswered;
-import org.apache.james.imap.api.message.request.UnDeleted;
-import org.apache.james.imap.api.message.request.UnDraft;
-import org.apache.james.imap.api.message.request.UnFlagged;
-import org.apache.james.imap.api.message.request.UnKeyword;
-import org.apache.james.imap.api.message.request.UnSeen;
 import org.apache.james.imap.api.message.response.StatusResponse;
 import org.apache.james.imap.api.message.response.StatusResponseFactory;
 import org.apache.james.imap.api.process.ImapProcessor;
 import org.apache.james.imap.api.process.SelectedMailbox;
 import org.apache.james.imap.encode.FakeImapSession;
-import org.apache.james.imap.message.request.SearchRequest;
 import org.apache.james.imap.message.response.SearchResponse;
 import org.apache.james.mailbox.MailboxManager;
 import org.apache.james.mailbox.MailboxSession;
@@ -268,7 +229,7 @@ public class SearchProcessorTest {
     @Test
     public void testBEFORE() throws Exception {
         expectsGetSelectedMailbox();
-        check(org.apache.james.imap.api.message.request.Before.apply(DAY_MONTH_YEAR), SearchQuery
+        check(org.apache.james.imap.search.Before.apply(DAY_MONTH_YEAR), SearchQuery
                 .internalDateBefore(getDate(DAY, MONTH, YEAR), DateResolution.Day));
     }
